old-api/api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
litellm
openai
qdrant-client
pydantic
pydantic-settings
tiktoken==0.5.1
